pyproject.toml
[tool.poetry]
name = "networksecurity"
version = "0.0.1"
description = "Network Security Project For Phishing Data"
authors = ["Darshan  <darshan922000@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
pandas = "^2.2.3"
numpy = "^2.2.1"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
certifi = "^2024.12.14"
scikit-learn = "^1.6.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
